Move arch specific tests
