+++ tests/CMakeLists.txt
@@ -26,11 +26,11 @@ if(MASTER_PROJECT)
 endif()
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/device)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/baremetal)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/blackhole)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/arch/blackhole)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/misc)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/pcie)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/unified)
-add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/wormhole)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/arch/wormhole)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/galaxy)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/test_utils)
