+++ tests/arch/blackhole/test_chip_info_bh.cpp
