+++ tests/arch/wormhole/test_arc_telemetry_wh.cpp
