+++ tests/galaxy/test_umd_concurrent_threads.cpp
@@ -21,7 +21,7 @@
 #include "tests/test_utils/fetch_local_files.hpp"
 #include "tests/test_utils/setup_risc_cores.hpp"
 #include "tests/test_utils/test_api_common.hpp"
-#include "tests/wormhole/test_wh_common.hpp"
+#include "tests/arch/wormhole/test_wh_common.hpp"
 #include "umd/device/cluster.hpp"
 #include "umd/device/cluster_descriptor.hpp"
 #include "wormhole/eth_interface.h"
